types: spread into multiple files for better separation, all exported from index
